third_party/blink/web_tests/http/tests/origin_trials/webexposed/compute-pressure-origin-trial-interfaces.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 ComputePressure_v2 --expire-timestamp=2000000000
-- -->
<meta http-equiv="origin-trial" content="AxZRvcj1/L8T6dR206rpmTEvEHFhCoYeVIEjYHvkkSdURAuX019bib6Rf0GJjwb6uh4OgjevQNBmQl7De5KJawsAAABaeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiQ29tcHV0ZVByZXNzdXJlX3YyIiwgImV4cGlyeSI6IDIwMDAwMDAwMDB9" />
<title>Compute Pressure API - interfaces exposed by origin trial</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/origin-trials-helper.js"></script>
<script>
test(t => {
  OriginTrialsHelper.check_properties_exist(this, {
    'PressureObserver': ['observe', 'unobserve', 'disconnect', 'takeRecords'],
  });
}, 'Compute Pressure API interfaces and properties in Origin-Trial enabled document.');

test(t => {
  assert_true('PressureObserver' in window, 'PressureObserver object exists on window scope');
  assert_not_equals(window.PressureObserver, undefined, 'PressureObserver can be accessed on window');
}, 'Compute Pressure API entry point in Origin-Trial enabled document.');
</script>
